[feat] MyShell, MyShellList 수정
