--- client/src/component/myshells/MyShellsList.styled.ts
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  max-height: 700px;
-  gap: 20px;
-  width: auto;
-  overflow-y: scroll;
-  padding: 5px;
-`;
-
-export { Wrapper };
